requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
abseil-cpp=20211102.0=h96cf925_1
aiohttp=3.8.1=py39h63b48b0_1
aiosignal=1.2.0=pyhd8ed1ab_0
aom=3.4.0=hb486fe8_1
appdirs=1.4.4=pyh9f0ad1d_0
async-timeout=4.0.2=pyhd8ed1ab_0
attrs=21.4.0=pyhd8ed1ab_0
audioread=2.1.9=py39h6e9494a_3
brotli=1.0.9=h5eb16cf_7
brotli-bin=1.0.9=h5eb16cf_7
brotlipy=0.7.0=py39h63b48b0_1004
bzip2=1.0.8=h0d85af4_4
c-ares=1.18.1=h0d85af4_0
ca-certificates=2022.6.15=h033912b_0
cachetools=5.0.0=pyhd8ed1ab_0
certifi=2022.6.15=py39h6e9494a_0
cffi=1.15.1=py39hae9ecf2_0
charset-normalizer=2.1.0=pyhd8ed1ab_0
cryptography=37.0.1=py39hf6deb26_0
cycler=0.11.0=pyhd8ed1ab_0
decorator=5.1.1=pyhd8ed1ab_0
ffmpeg=5.0.1=gpl_h0bbc480_105
fonttools=4.33.3=py39h701faf5_0
freetype=2.10.4=h4cff582_1
frozenlist=1.3.0=py39h63b48b0_1
gettext=0.19.8.1=hd1a6beb_1008
giflib=5.2.1=hbcb3906_2
gmp=6.2.1=h2e338ed_0
gnutls=3.7.6=h047ecf1_5
google-api-core=2.8.2=pyhd8ed1ab_0
google-api-core-grpc=2.8.2=hd8ed1ab_0
google-auth=2.9.0=pyh6c4a22f_0
google-cloud-vision=2.7.3=pyhd8ed1ab_0
googleapis-common-protos=1.56.3=py39h6e9494a_0
grpc-cpp=1.47.0=h6f0e29b_1
grpcio=1.47.0=py39h35003fd_1
grpcio-status=1.46.3=pyhd8ed1ab_0
icu=70.1=h96cf925_0
idna=3.3=pyhd8ed1ab_0
joblib=1.1.0=pyhd8ed1ab_0
jpeg=9e=hac89ed1_2
kiwisolver=1.4.3=py39h7c694c3_0
krb5=1.19.3=hb98e516_0
lame=3.100=h35c211d_1001
lcms2=2.12=h577c468_0
lerc=3.0=he49afe7_0
libblas=3.9.0=15_osx64_openblas
libbrotlicommon=1.0.9=h5eb16cf_7
libbrotlidec=1.0.9=h5eb16cf_7
libbrotlienc=1.0.9=h5eb16cf_7
libcblas=3.9.0=15_osx64_openblas
libcxx=14.0.6=hce7ea42_0
libdeflate=1.12=hac89ed1_0
libedit=3.1.20191231=h0678c8f_2
libffi=3.4.2=h0d85af4_5
libflac=1.3.4=h96cf925_0
libgfortran=5.0.0=9_3_0_h6c81a4c_23
libgfortran5=9.3.0=h6c81a4c_23
libiconv=1.16=haf1e3a3_0
libidn2=2.3.2=h0d85af4_0
liblapack=3.9.0=15_osx64_openblas
libllvm11=11.1.0=hd011deb_3
libogg=1.3.4=h35c211d_1
libopenblas=0.3.20=openmp_hb3cd9ec_0
libopus=1.3.1=hc929b4f_1
libpng=1.6.37=h7cec526_2
libpq=14.4=h2b7167c_0
libprotobuf=3.20.1=h2292cb8_0
librosa=0.9.2=pyhd8ed1ab_0
libsndfile=1.0.31=he49afe7_1
libtasn1=4.18.0=hac89ed1_1
libtiff=4.4.0=h9847915_1
libunistring=0.9.10=h0d85af4_0
libvorbis=1.3.7=h046ec9c_0
libvpx=1.11.0=he49afe7_3
libwebp=1.2.2=h28dabe5_0
libwebp-base=1.2.2=h0d85af4_1
libxcb=1.13=h0d85af4_1004
libxml2=2.9.14=h08a9926_1
libzlib=1.2.12=hfe4f2af_1
llvm-openmp=14.0.4=ha654fa7_0
llvmlite=0.38.1=py39hdfe08e5_0
lz4-c=1.9.3=he49afe7_1
matplotlib-base=3.5.2=py39h64a0072_0
multidict=6.0.2=py39h63b48b0_1
munkres=1.1.4=pyh9f0ad1d_0
ncurses=6.3=h96cf925_1
nettle=3.8=h96f3785_0
numba=0.55.2=py39hc37845d_0
numpy=1.22.4=py39h677350a_0
openh264=2.2.0=h2292cb8_0
openjpeg=2.4.0=h6e7aa92_1
openssl=3.0.4=hfe4f2af_2
p11-kit=0.24.1=h65f8906_0
packaging=21.3=pyhd8ed1ab_0
pillow=9.1.1=py39h579eac4_1
pip=22.1.2=pyhd8ed1ab_0
pooch=1.6.0=pyhd8ed1ab_0
proto-plus=1.20.6=pyhd8ed1ab_0
protobuf=3.20.1=py39hd408605_0
psycopg2=2.9.3=py39hdb5d3f9_0
pthread-stubs=0.4=hc929b4f_1001
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.21=pyhd8ed1ab_0
pydub=0.25.1=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.9=pyhd8ed1ab_0
pysocks=1.7.1=py39h6e9494a_5
pysoundfile=0.10.3.post1=pyhd3deb0d_0
python=3.9.7=h38b4d05_3_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pyu2f=0.1.5=pyhd8ed1ab_0
re2=2022.06.01=hb486fe8_0
readline=8.1.2=h3899abd_0
requests=2.28.1=pyhd8ed1ab_0
requests-toolbelt=0.9.1=py_0
resampy=0.3.0=pyhd8ed1ab_0
rsa=4.8=pyhd8ed1ab_0
scikit-learn=1.1.1=py39hafe9688_0
scipy=1.8.1=py39hfa1a3ab_0
setuptools=62.6.0=py39h6e9494a_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.39.0=hd9f0692_0
svt-av1=1.1.0=hb486fe8_1
threadpoolctl=3.1.0=pyh8a188c0_0
tk=8.6.12=h5dbffcc_0
typing-extensions=4.3.0=hd8ed1ab_0
typing_extensions=4.3.0=pyha770c72_0
tzdata=2022a=h191b570_0
unicodedata2=14.0.0=py39h63b48b0_1
urllib3=1.26.9=pyhd8ed1ab_0
wheel=0.37.1=pyhd8ed1ab_0
x264=1!161.3030=h0d85af4_1
x265=3.5=hbb4e6a2_3
xorg-libxau=1.0.9=h35c211d_0
xorg-libxdmcp=1.1.3=h35c211d_0
xz=5.2.5=haf1e3a3_1
yarl=1.7.2=py39h63b48b0_2
zlib=1.2.12=hfe4f2af_1
zstd=1.5.2=ha9df2e0_2
